+++ src/backend/cdb/cdbpartition.c
@@ -38,6 +38,7 @@
 #include "parser/analyze.h"
 #include "parser/parse_expr.h"
 #include "parser/parse_oper.h"
+#include "parser/parse_partition.h"
 #include "parser/parse_relation.h"
 #include "parser/parse_target.h"
 #include "tcop/utility.h"
